+++ R/likelihood.R
@@ -1,6 +1,12 @@
 #' @title Initialize a Poisson likelihood node
 #'
-#' @description This function constructs a likelihood object to represent a Poisson likelihood node.
+#' @description
+#'
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This function is deprecated and will be removed in the next release.
+#'     To incorporate a Poisson likelihood node into a DGP structure,
+#'     use the `likelihood` argument in the `dgp()` function instead.
 #'
 #' @param input_dim a vector of length one that contains the indices of one GP node in the feeding
 #'     layer whose outputs feed into this likelihood node. When set to `NULL`,
@@ -18,12 +24,23 @@
 #' # on how to customize DGP structures using Poisson().
 #' }
 #' @md
+#' @keywords internal
 #' @export
 Poisson <- function(input_dim = NULL) {
   if ( is.null(pkg.env$dgpsi) ) {
     init_py(verb = F)
     if (pkg.env$restart) return(invisible(NULL))
   }
+
+  lifecycle::deprecate_warn(
+    when = "3.0.0",
+    what = "kernel()",
+    details = c(i = "The function will be removed in the next release.",
+                i=  "It may not be compatible with other functions in this version.",
+                i = "Please use the `likelihood` argument in `dgp()` function to incorporate a Poisson likelihood node into a DGP structure."
+    )
+  )
+
   if(!is.null(input_dim)){
     input_dim <- reticulate::np_array(as.integer(input_dim - 1))
   }
@@ -34,7 +51,13 @@ Poisson <- function(input_dim = NULL) {
 
 #' @title Initialize a heteroskedastic Gaussian likelihood node
 #'
-#' @description This function constructs a likelihood object to represent a heteroskedastic Gaussian likelihood node.
+#' @description
+#'
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This function is deprecated and will be removed in the next release.
+#'     To incorporate a heteroskedastic Gaussian likelihood node into a DGP structure,
+#'     use the `likelihood` argument in the `dgp()` function instead.
 #'
 #' @param input_dim a vector of length two that contains the indices of two GP nodes in the feeding
 #'     layer whose outputs feed into this likelihood node. When set to `NULL`,
@@ -52,12 +75,23 @@ Poisson <- function(input_dim = NULL) {
 #' # on how to customize DGP structures using Hetero().
 #' }
 #' @md
+#' @keywords internal
 #' @export
 Hetero <- function(input_dim = NULL) {
   if ( is.null(pkg.env$dgpsi) ) {
     init_py(verb = F)
     if (pkg.env$restart) return(invisible(NULL))
   }
+
+  lifecycle::deprecate_warn(
+    when = "3.0.0",
+    what = "kernel()",
+    details = c(i = "The function will be removed in the next release.",
+                i=  "It may not be compatible with other functions in this version.",
+                i = "Please use the `likelihood` argument in `dgp()` function to incorporate a heteroskedastic Gaussian likelihood node into a DGP structure."
+    )
+  )
+
   if(!is.null(input_dim)){
     input_dim <- reticulate::np_array(as.integer(input_dim - 1))
   }
@@ -67,7 +101,13 @@ Hetero <- function(input_dim = NULL) {
 
 #' @title Initialize a negative Binomial likelihood node
 #'
-#' @description This function constructs a likelihood object to represent a negative Binomial likelihood node.
+#' @description
+#'
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This function is deprecated and will be removed in the next release.
+#'     To incorporate a negative Binomial likelihood node into a DGP structure,
+#'     use the `likelihood` argument in the `dgp()` function instead.
 #'
 #' @param input_dim a vector of length two that contains the indices of two GP nodes in the feeding
 #'     layer whose outputs feed into this likelihood node. When set to `NULL`,
@@ -85,58 +125,26 @@ Hetero <- function(input_dim = NULL) {
 #' # on how to customize DGP structures using NegBin().
 #' }
 #' @md
+#' @keywords internal
 #' @export
 NegBin <- function(input_dim = NULL) {
   if ( is.null(pkg.env$dgpsi) ) {
     init_py(verb = F)
     if (pkg.env$restart) return(invisible(NULL))
   }
+
+  lifecycle::deprecate_warn(
+    when = "3.0.0",
+    what = "kernel()",
+    details = c(i = "The function will be removed in the next release.",
+                i=  "It may not be compatible with other functions in this version.",
+                i = "Please use the `likelihood` argument in `dgp()` function to incorporate a negative Binomial likelihood node into a DGP structure."
+    )
+  )
+
   if(!is.null(input_dim)){
     input_dim <- reticulate::np_array(as.integer(input_dim - 1))
   }
   res <- pkg.env$dgpsi$NegBin(input_dim)
-  return(res)
-}
-
-#' @title Initialize a Categorical likelihood node
-#'
-#' @description
-#'
-#' `r new_badge("new")`
-#'
-#' This function constructs a likelihood object to represent a categorical likelihood node.
-#'
-#' @param K an integer specifying the number of classes in the training data. If `K` is set to `NULL` (the default),
-#' its value will be inferred from the training data when running [dgp()]. Defaults to `NULL`.
-#' @param input_dim a vector that specifies which GP node outputs feed into this likelihood node. It should be either:
-#' * A vector of length `1` (for binary output), or
-#' * A vector of length `K` (for `K`-class output), containing the indices of one or `K` GP nodes in the feeding layer.
-#'
-#' If `input_dim` is set to `NULL` (the default), all GP node outputs from the feeding layer feed into the likelihood node.
-#' In this case, you must ensure that the feeding layer contains exactly one or `K` GP nodes. Defaults to `NULL`.
-#'
-#' @return A 'python' object to represent a categorical likelihood node.
-#' @note The categorical likelihood node can only be linked to one or `K` feeding GP nodes.
-#' @details See further examples and tutorials at <https://mingdeyu.github.io/dgpsi-R/>.
-#' @examples
-#' \dontrun{
-#'
-#' # Check https://mingdeyu.github.io/dgpsi-R/ for examples
-#' # on how to customize DGP structures using Categorical().
-#' }
-#' @md
-#' @export
-Categorical <- function(K = NULL, input_dim = NULL) {
-  if ( is.null(pkg.env$dgpsi) ) {
-    init_py(verb = F)
-    if (pkg.env$restart) return(invisible(NULL))
-  }
-  if(!is.null(input_dim)){
-    input_dim <- reticulate::np_array(as.integer(input_dim - 1))
-  }
-  if(!is.null(K)){
-    K <- reticulate::np_array(as.integer(K))
-  }
-  res <- pkg.env$dgpsi$Categorical(K, input_dim)
   return(res)
 }